Move pkg/remote to server/registry/test/remote
