+++ server/registry/test/remote/proxy.go
